refactor(agents): move prompt-audit ownership from adapter to AgentManager (#523)
AgentManager is the universal chokepoint for every LLM call (ADR-013 §0,
compiler-enforced). SessionManager.runInSession() only covers descriptor
sessions (Path A); the ~19 direct agentManager.run/complete call sites in
review, debate, rectification, and CLI were never reachable via
auditCallback injection.

Changes:
- src/agents/audit.ts: canonical audit writer (moved from acp/prompt-audit.ts)
  Adds agentName and stableSessionId fields; makes sessionName optional with
  agentName fallback in filename; per-turn granularity removed (one entry per
  AgentManager.run/complete call — tool-continuation prompts are adapter-internal
  plumbing, not distinct LLM calls).
- src/agents/acp/prompt-audit.ts: thin re-export shim (preserves test imports)
- src/agents/acp/adapter.ts: remove writePromptAudit import and both audit blocks;
  drop unused sessionResumed and config locals
- src/agents/manager.ts: fire writePromptAudit in runWithFallback() and
  completeWithFallback() before delegating to adapter
- src/agents/types.ts: add projectDir to CompleteOptions (needed for audit path)
- src/agents/index.ts: export audit API from agents barrel

Closes #523.
